Make tests check if project exists before trying to delete it, preventing exceptions from reaching the logs
SVN r9991 |2008-12-22 20:55:57 +0000

+++ src/org/labkey/test/bvt/AuditLogTest.java
@@ -44,7 +44,10 @@ public class AuditLogTest extends BaseSeleniumWebTest
     protected void doCleanup() throws Exception
     {
         deleteUser(AUDIT_TEST_USER);
-        deleteProject(AUDIT_TEST_PROJECT);
+        if (isLinkPresentContainingText(AUDIT_TEST_PROJECT))
+        {
+            deleteProject(AUDIT_TEST_PROJECT);
+        }
     }
 
     protected void doTestSteps() throws Exception
